+++ src/components/media/Badge.jsx
@@ -4,15 +4,11 @@ import cx from 'classnames';
 
 export const SIZES = ['xs', 's'];
 
-const Badge = ({ label, size, ...rest }) => (
+const Badge = ({ label, size = 's', ...rest }) => (
   <div {...rest} className={cx('fdsBadge', 'wrap--singleLine', `fdsBadge--${size}`)}>
     {label}
   </div>
 );
-
-Badge.defaultProps = {
-  size: 's',
-};
 
 Badge.propTypes = {
   /** Text inside the button */