+++ src/components/popovers/MenuItem.jsx
@@ -2,8 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { MenuItem as ReachMenuItem } from '@reach/menu-button';
 
-const MenuItem = ({ onSelect, children }) => (
-  <ReachMenuItem onSelect={onSelect}>{children}</ReachMenuItem>
+const MenuItem = ({ onSelect, children, isDisabled }) => (
+  <ReachMenuItem
+    className={isDisabled && 'fdsMenuItem--disabled'}
+    onSelect={!isDisabled && onSelect}
+    aria-disabled={isDisabled}
+  >
+    {children}
+  </ReachMenuItem>
 );
 
 MenuItem.propTypes = {
@@ -11,7 +17,11 @@ MenuItem.propTypes = {
   onSelect: PropTypes.func.isRequired,
 
   /** React children - any JSX can be used in a menu item */
-  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]).isRequired,
+  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node])
+    .isRequired,
+
+  /** Styles menu item as disabled and disables selection events */
+  isDisabled: PropTypes.bool,
 };
 
 export default MenuItem;